Make SQLite database available to all tests
